+++ CHANGELOG.md
@@ -25,6 +25,7 @@ Please document your changes in this format:
 - Rename stores to places in UI and change shopping-cart to dot-circle #1398 @tiltec @djahnie
 - Separate conversations and threads in message notifications @tiltec @djahnie
 - Move language picker to user settings @tiltec @djahnie
+- Move community feed to sidenav @tiltec @djahnie
 
 ### Fixed
 - Permission checks in main routes had a typo and were not active @djahnie @tiltec